requirements-pkg.txt
django_admin_report==0.4
django-import-export==0.5.1
django-oauth-toolkit==1.0.0
django-simple-history==1.9.0
elasticsearch==5.4.0
factory_boy==2.9.2
